FLTK widget defers update_image() calls until draw()
If the GL context wasn't inited yet at update_image() time, I must init it
first. BUT in order to init it, some things about the X window must be set up. I
cannot rely on them being set up here, so I init stuff only in the draw() call
below. If I try to init here, I see this:

  python3: ../src/dispatch_common.c:868: epoxy_get_proc_address: Assertion `0 && "Couldn't find current GLX or EGL context.\n"' failed.

So I save the data in this call, and apply it later, when I'm ready