FLTK widget: update_image is NOT virtual
This is a revert of

  commit 8b1b42bcda23f711be62a512fba0df7f99d6cac4
  Author: Dima Kogan <dima@secretsauce.net>
  Date:   Wed Mar 23 14:32:39 2022 -0700
      FLTK widget: update_image() is virtual

That commit was incomplete. update_image() uses a typemap(in) for the
swig-wrapping. If this function is virtual, it ALSO needs a director-specific
typemap to convert the python arguments back to C++ ones. I'm not going to do
that just yet, so I simply revert the commit for now.