Fl_Gl_Image_Widget constructor: double-buffering can be turned on/off
On some hardware (i915 for instance) double-buffering causes redrawing, so
turning it off is a good workaround. In general, single-buffering causes redraw
flicker, so double-buffering is recommended where possible.

Currently double-buffering is the default, which is DIFFERENT from the previous
default: single-buffering.

This applies to C++ and Python.

THIS IS AN ABI BREAK, but not an API break