Reverted "Th Fl_Gl_Image_Widget can't appear focused"
This commit:

  commit 3ec270f36a39da6ea4a0380a6dd22cf8b593ec76
  Author: Dima Kogan <dima@secretsauce.net>
  Date:   Sat Aug 24 12:59:49 2024 -0700

    Th Fl_Gl_Image_Widget can't appear focused

    and shouldn't be a part of navigation commands. See:
    https://github.com/fltk/fltk/discussions/1048

The intent was to prevent the navigation (left/right/up/down) keyboard commands
to be implicitly picked up the Fl_Gl_Image_Widget, so that other widgets could
get a chance to process these. This worked, but had a secondary effect that no
keys at all could be processed by this widget. The "right" solution probably is
to do the appropriate thing in a parent class, as described in the discussion
link above. I can't do this inside Fl_Gl_Image_Widget, so I'm going to leave
this to the user.